# docs/swagger.yaml
definitions:
  dto.DeleteTaskResponse:
    properties:
      success:
        example: true
        type: boolean
    required:
    - success
    type: object
  dto.ShowTaskResponse:
    properties:
      address:
        example: TU...123
        type: string
      currency:
        example: TRX
        type: string
      description:
        example: Take a several photo of a car
        type: string
      id:
        example: 01a09bc4-29dc-7cb3-8966-93d058554dbb
        type: string
      message:
        example: Please make the payment within 1 hour
        type: string
      price:
        example: "10.0"
        type: string
      status:
        example: PENDING_PAYMENT
        type: string
      type:
        example: TAKE_PHOTO
        type: string
    required:
    - address
    - currency
    - description
    - id
    - price
    - status
    - type
    type: object
  httpd.InternalError:
    properties:
      code:
        example: "500"
        type: string
      message:
        example: Internal error
        type: string
    required:
    - code
    - message
    type: object
  httpd.NotFound:
    properties:
      code:
        example: "404"
        type: string
      message:
        example: Not found
        type: string
    required:
    - code
    - message
    type: object
  httpd.ValidationError:
    properties:
      code:
        example: "422"
        type: string
      message:
        example: Validation error
        type: string
    required:
    - code
    - message
    type: object
info:
  contact: {}
paths:
  /tasks:
    get:
      consumes:
      - application/json
      description: Get the list of the tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ShowTaskResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpd.InternalError'
      summary: Show the list of the task
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: 'To create a new task, you must provide the mandatory fields: description,
        type, price, and currency.'
      parameters:
      - description: Detailed task description for the person performing the work
        in: body
        name: description
        required: true
        schema:
          type: string
      - description: Type of the task
        in: body
        name: type
        required: true
        schema:
          enum:
          - TAKE_PHOTO
          - TAKE_VIDEO
          - OTHER
          type: string
      - description: Price of the task
        in: body
        name: price
        required: true
        schema:
          type: string
      - description: Currency of the price
        in: body
        name: currency
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ShowTaskResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpd.ValidationError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpd.InternalError'
      summary: Create a task
      tags:
      - tasks
  /tasks/{guid}:
    delete:
      consumes:
      - application/json
      description: Delete the task by ID
      parameters:
      - description: Task GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dto.DeleteTaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpd.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpd.InternalError'
      summary: Delete the task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get task info by ID
      parameters:
      - description: Task ID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ShowTaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpd.NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpd.InternalError'
      summary: Show a task
      tags:
      - tasks
swagger: "2.0"